backend/requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
auto-py-to-exe==2.9.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.3
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.11
django-allauth==0.44.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-mssql-backend==2.8.1
django-rest-auth==0.9.5
django-rest-multiple-models==2.1.3
djangorestframework==3.11.1
Eel==0.12.4
filelock==3.0.12
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pefile==2021.5.24
psycopg2==2.8.5
pycparser==2.20
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
PyJWT==2.0.0
pylint==2.6.0
pyodbc==4.0.30
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python3-openid==3.2.0
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.2
virtualenv==20.2.1
virtualenvwrapper-win==1.2.6
whichcraft==0.6.1
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
